Move CFamilyHelper to CFamily namespace
